requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.2.0